back/requirements.txt
Django
gunicorn
psycopg2-binary
djangorestframework
djangorestframework-jsonapi
django-cors-headers
